+++ src/api/helpers.rs
@@ -58,6 +58,8 @@ pub fn parse_user_details(
         }
     }
     
+    // This probably works for the most part. It may have issues on certain strings so it's excluded
+    // for now.
     // fn get_persona_names(contents: &str) -> Result<(String, String), ParseHtmlError> {
     //     let my_persona_name = regex_captures!(r#"var g_strYourPersonaName = "(?:[^"\\]|\\.)*";\n"#, contents)
     //         .map(|(_, name)| unescape(name))